+++ lib/plug_locale/web_browser/config.ex
@@ -5,7 +5,7 @@ defmodule PlugLocale.WebBrowser.Config do
     :default_locale,
     :locales,
     :detect_locale_from,
-    :sanitize_locale_by,
+    :cast_locale_by,
     :route_identifier,
     :path_param_key,
     :assign_key,
@@ -21,7 +21,7 @@ defmodule PlugLocale.WebBrowser.Config do
     detect_locale_from =
       Keyword.get(opts, :detect_locale_from, [:cookie, :referrer, :accept_language])
 
-    sanitize_locale_by = Keyword.get(opts, :sanitize_locale_by, &PlugLocale.Sanitizer.sanitize/1)
+    cast_locale_by = Keyword.get(opts, :cast_locale_by, nil)
     route_identifier = Keyword.get(opts, :route_identifier, :locale)
     assign_key = Keyword.get(opts, :assign_key, route_identifier)
     query_key = Keyword.get(opts, :query_key, to_string(route_identifier))
@@ -31,13 +31,12 @@ defmodule PlugLocale.WebBrowser.Config do
       default_locale: default_locale,
       locales: locales,
       detect_locale_from: detect_locale_from,
-      sanitize_locale_by: sanitize_locale_by,
+      cast_locale_by: cast_locale_by,
       route_identifier: route_identifier,
       assign_key: assign_key,
       query_key: query_key,
       cookie_key: cookie_key
     ]
-    |> validate!()
     |> as_map!()
     |> as_struct!()
   end
@@ -57,7 +56,7 @@ defmodule PlugLocale.WebBrowser.Config do
     default_locale = Keyword.fetch!(opts, :default_locale)
     locales = Keyword.fetch!(opts, :locales)
     detect_locale_from = Keyword.fetch!(opts, :detect_locale_from)
-    sanitize_locale_by = Keyword.fetch!(opts, :sanitize_locale_by)
+    cast_locale_by = Keyword.fetch!(opts, :cast_locale_by)
     route_identifier = Keyword.fetch!(opts, :route_identifier)
     assign_key = Keyword.fetch!(opts, :assign_key)
     query_key = Keyword.fetch!(opts, :query_key)
@@ -67,7 +66,7 @@ defmodule PlugLocale.WebBrowser.Config do
       default_locale: default_locale,
       locales: Enum.uniq([default_locale | locales]),
       detect_locale_from: detect_locale_from,
-      sanitize_locale_by: sanitize_locale_by,
+      cast_locale_by: cast_locale_by,
       route_identifier: route_identifier,
       path_param_key: to_string(route_identifier),
       assign_key: assign_key,