+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
+## Unreleased
+
+Breaking changes:
+
+- rename `:sanitize_locale_by` option `:cast_locale_by`
+
 ## v0.3.0
 
 Breaking changes:
 
-- `:sanitize_locale` is renamed to `:sanitize_locale_by`
+- rename `:sanitize_locale` option `:sanitize_locale_by`